% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWadmiX.R
\name{HWadmiX}
\alias{HWadmiX}
\title{Carries out likelihood ratio test and chi-squared test for Hardy-Weinberg equilibrium for 
X chromosomal markers, accounting for the possibility of sex-biased admixture}
\usage{
HWadmiX(X)
}
\arguments{
\item{X}{a named vector containing the genotype counts (names should be AA, AB, BB, A, B)}
}
\value{
list with the following elements:

  lrt_pval: the p-value of the LRT test
  
  prev_p_f, prev_p_m: the ML estimates of the previous generation allele frequencies in females and males
  
  F_ic: the ML estimate of the inbreeding coefficient
  
  current_p_f, current_p_m: estimates of the current generation allele frequencies in females and males, 
  based on the ML estimates of the previous generation allele frequencies and of the inbreeding coefficient
  
  current_p_AA, current_p_AB, current_p_BB: estimates of the current generation genotype frequencies in females, 
  based on the ML estimates of the previous generation allele frequencies and of the inbreeding coefficient
  
  cs_ml_pval: the p-value of the chi-squared (maximum likelihood) test
  
  you_lrt0_unconstrained_pval: the p-value of the LRT0 test from 
       X.-P. You, Q.-L. Zou, J.-L. Li and J.-Y. Zhou, Likelihood ratio test for 
       excess homozygosity at marker loci on X chromosome. LRT0 test, PLoS One 2015; 10(12): e0145032,
       where the inbreeding coefficient is not constrained to be in [0,1]
}
\description{
Carries out likelihood ratio test and chi-squared test for Hardy-Weinberg equilibrium for 
X chromosomal markers, accounting for the possibility of sex-biased admixture
}
\examples{
snp <- c(A = 122, B = 278, AA = 88, AB = 242, BB = 70)
HWadmiX(snp)

}

