% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWadmix.R
\name{HWadmix}
\alias{HWadmix}
\title{Carries out likelihood ratio test and chi-squared testfor Hardy Weinberg equilibrium for 
X chromosomal markers, accounting for the possibility of sex-biased admixture}
\usage{
HWadmix(X)
}
\arguments{
\item{X}{a named vector containing the genotype counts (names should be AA, AB, BB, A, B)}
}
\value{
list with elements
  lrt_pval: the p-value of the LRT test
  prev_p_f, prev_p_m: the ML estimates of the previous generation allele frequencies in females and males
  F_ic: the ML estimate of the inbreeding coefficient
  current_p_f, current_p_m: the ML estimates of the current generation allele frequencies in females and males
  current_p_AA, current_p_AB, current_p_BB: the ML estimates of the current generation genotype frequencies in females
  cs_ml_pval: the p-value of the chi-squared (maximum likelihood) test
  you_lrt0_unconstrained_pval: the p-value of the You et al. LRT0 test
       where the inbreeding coefficient is not constrained to be in [0,1]
}
\description{
Carries out likelihood ratio test and chi-squared testfor Hardy Weinberg equilibrium for 
X chromosomal markers, accounting for the possibility of sex-biased admixture
}
\examples{
snp <- c(A = 122, B = 278, AA = 88, AB = 242, BB = 70)
HWadmix(snp)

}

